point changed from double to int

+++ Translator/Translator/Header files/point.h
@@ -4,9 +4,9 @@ using namespace std;
 
 class Point {
 public:
-    double x, y, z;  
+    int x, y, z;  
 
-    Point(double x = 0, double y = 0, double z = 0);  
+    Point(int x = 0, int y = 0, int z = 0);  
     ~Point();  
 
     void get_coord() const;
